feat[next]: Add `FieldBuiltinFuncRegistry` mixin (#1330)
Adds FieldBuiltinFuncRegistry to allow Field subclasses to register their own builtins
